db_service/requirements.txt
fastapi
uvicorn
psycopg2-binary
asyncpg
sqlalchemy>=2.0
uuid-utils>=0.11.0  # provides uuid7() support until native uuid.UUIDv7 is available in Python 3.14+
argon2-cffi>=23.1.0
pydantic[email]
pandas
# pandas=2.2.3
apscheduler
httpx
